class User modif

+++ grails-app/domain/alloserie/User.groovy
@@ -13,7 +13,7 @@ class User {
     static constraints = {
       firstName(blank:false , maxSize: 20, nullable: false )
       lastName(blank:false , maxSize: 20, nullable: false)
-      age(nullable: false , max: 105, min: 5)
+      age(nullable: false , max: (short)105, min:  (short)5)
       logging(nullable: false , maxSize: 20, minSize: 5, unique: true)
       password(nullable: false, minSize: 6)
       userType(inList: "Admin", "User")
